feat: Add TrainerTrainingInfoResponse DTO.
Introduces the `TrainerTrainingInfoResponse` Data Transfer Object (DTO).
This DTO is designed to present detailed information about a training session
from the perspective of a Trainer. It includes training name, date, type, duration,
and the name of the associated trainee. This DTO is specifically used for listing
a trainer's training history.